menuto-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
openai==1.3.5
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
pydantic==2.5.0
python-dotenv==1.0.0
PyMuPDF==1.23.8